feat - add abstract class
